style: clean up redux middleware
