+++ renderer/store/middleware/index.js
@@ -1,0 +1,2 @@
+export debounceMiddleware from './debounceMiddleware'
+export userTimingMiddleware from './userTimingMiddleware'
